+++ Craftsman/Models/BffEntityTemplate.cs
@@ -1,0 +1,10 @@
+namespace Craftsman.Models;
+
+using Enums;
+using Exceptions;
+using Helpers;
+
+public class BffEntityTemplate : BffBased
+{
+    public List<BffEntity> Entities { get; set; }
+}