+++ Craftsman/Models/BffTemplate.cs
@@ -1,18 +1,9 @@
 namespace Craftsman.Models;
 
-using Enums;
-using Exceptions;
 using Helpers;
 
-public class BffTemplate
+public class BffTemplate : BffBased
 {
-    private string _projectName;
-    public string ProjectName
-    {
-        get => _projectName;
-        set => _projectName = value ?? throw new Exception("Project name is required for a BFF template.");
-    }
-
     private string _profileName;
     public string ProfileName
     {
@@ -69,108 +60,4 @@ public class BffTemplate
     }
 
     public List<BffEntity> Entities { get; set; }
-}
-
-public class BffEntity
-{
-    private string _name;
-    /// <summary>
-    /// The name of the entity
-    /// </summary>
-    public string Name
-    {
-        get => _name.UppercaseFirstLetter();
-        set => _name = value;
-    }
-
-    private string _plural;
-    /// <summary>
-    /// The Plural name of the entity.
-    /// </summary>
-    public string Plural
-    {
-        get => _plural ?? $"{Name}s";
-        set => _plural = value;
-    }
-
-    /// <summary>
-    /// List of properties associated to the entity
-    /// </summary>
-    public List<BffEntityProperty> Properties { get; set; } = new List<BffEntityProperty>();
-    
-    private List<BffFeature> _features = new List<BffFeature>();
-    public List<BffFeature> Features
-    {
-        get => _features;
-        set
-        {
-            var entitylessFeatures = value ?? new List<BffFeature>();
-            _features = entitylessFeatures
-                .Select(f =>
-                {
-                    f.EntityName = Name;
-                    f.EntityPlural = Plural;
-                    return f;
-                })
-                .ToList();
-        }
-    }
-}
-
-public class BffFeature
-{
-    public string EntityName { get; set; }
-
-    public string EntityPlural { get; set; }
-        
-    private FeatureType FeatureType { get; set; }
-    public string Type
-    {
-        get => FeatureType.Name;
-        set
-        {
-            if (!FeatureType.TryFromName(value, true, out var parsed))
-            {
-                if(value.Equals("CreateRecord", StringComparison.InvariantCultureIgnoreCase))
-                    FeatureType = FeatureType.AddRecord;
-                else
-                    throw new InvalidFeatureTypeException(value);
-            }
-            FeatureType = parsed;
-        }
-    }
-
-    public string BffApiName => FeatureType == FeatureType.GetList 
-            ? FeatureType.BffApiName(EntityPlural) 
-            : FeatureType.BffApiName(EntityName);
-}
-
-public class BffEntityProperty
-{
-    private string _name;
-    /// <summary>
-    /// Name of the property
-    /// </summary>
-    public string Name
-    {
-        get => _name.UppercaseFirstLetter();
-        set => _name = value;
-    }
-
-    private string _type = "string";
-    /// <summary>
-    /// Type of property (e.g. string, int, DateTime?, etc.)
-    /// </summary>
-    public string Type
-    {
-        get => _type;
-        set => _type = Utilities.PropTypeCleanupTypeScript(value);
-    }
-
-    public bool Nullable => Type.Contains('?');
-
-    /// <summary>
-    /// The Type with the optional marker removed
-    /// </summary>
-    public string RawType => Type.Replace("?", "");
 }