+++ CHANGELOG.md
@@ -46,6 +46,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   * Minor helper override on fake generators
   
   * New `add:bff` command to add a bff to your solution
+  
+  * New `add:bffentity` command to add entities to your bff
 
 
 ### Updated